Refactor Splade++
